third_party/WebKit/LayoutTests/fast/dom/DocumentType/previous-element-sibling.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<script>

description('This tests the ChildNode interface previousElementSibling property on a DocumentType node.');

var doctype = document.doctype;
shouldBeNull('doctype.previousElementSibling');
document.insertBefore(new Comment(), doctype);
shouldBeNull('doctype.previousElementSibling');

</script>
